QCamera2: Don't send dump API to statemachine.
Dump logic is not implemented in the current HAL version, so no
need to send to statemachine. In client death scenarios, dump
call to HAL is causing deadlock scenario due to some bug in
frameworks (camera service layer). Since dump logic is anyways
not implemented, removing the API send/wait for dump will avoid
the deadlock and serve as a workaround for the bug in frameworks.

Change-Id: I87882e0640a82239a2ab0b9ad006297c269c2b0b